==> notebooks/1.0-ak-extract-titanic-data.ipynb <==
## Extracting Titanic Disaster Data From Kaggle

#### MAKE SURE YOU HAVE MENTIONED YOUR KAGGLE USERNAME AND PASSWORD IN ".env" FILE AS DISCUSSED IN THE CLIP.

!conda install python-dotenv

from dotenv import load_dotenv, find_dotenv

# find .env automatically by walking up directories until it's found
dotenv_path = find_dotenv()
# load up the entries as environment variables
load_dotenv(dotenv_path)

# imports
import requests
from requests import session
import os
from dotenv import load_dotenv, find_dotenv
import csv

!kaggle competitions download -c titanic -f test.csv -p ../data/raw
!kaggle competitions download -c titanic -f train.csv -p ../data/raw

with open('../data/raw/train.csv', 'r') as csvFile:
    reader = csv.reader(csvFile)
    for row in reader:
        print(row)
csvFile.close()

!ls -l ../data/raw

### Builiding the file script

get_raw_data_script_file = os.path.join(os.path.pardir,'src','data','get_raw_data.py')

%%writefile $get_raw_data_script_file
# -*- coding: utf-8 -*-
import os
from dotenv import find_dotenv, load_dotenv
from requests import session
import logging

                
def main(project_dir):
    '''
    main method
    '''
    # get logger
    logger = logging.getLogger(__name__)
    logger.info('getting raw data')

    # extract data
    os.system('kaggle competitions download -c titanic -f test.csv -p ../data/raw')
    os.system('kaggle competitions download -c titanic -f train.csv -p ../data/raw')
    logger.info('downloaded raw training and test data')

if __name__ == '__main__':
    # getting root directory
    project_dir = os.path.join(os.path.dirname(__file__), os.pardir, os.pardir)
    
    # setup logger
    log_fmt = '%(asctime)s - %(name)s - %(levelname)s - %(message)s'
    logging.basicConfig(level=logging.INFO, format=log_fmt)

    # find .env automatically by walking up directories until it's found
    dotenv_path = find_dotenv()
    # load up the entries as environment variables
    load_dotenv(dotenv_path)

    # call the main
    main(project_dir)


!python $get_raw_data_script_file
